mask nicknames in IRC with unicode word joiners
Some uesrs who are connected to slack and IRC at the same time compained that the IRC-Part of the bridge mentions them every time they write anything in the slack. That's annoying. The solution is to insert an unicode word joiner after every character of the nickname. That causes the IRC-Client not to notice that their nick was mentioned supressing the annoying notification.